requirements.txt
# ── Core OpenEnv framework ──────────────────────────────────────────────────
openenv-core>=0.2.2

# ── Server dependencies ─────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0          # [standard] includes websockets + httptools
pydantic>=2.0.0

# ── WebSocket support (required for OpenEnv /ws endpoint) ───────────────────
websockets>=11.0

# ── Baseline inference script ────────────────────────────────────────────────
openai>=1.0.0

# ── HTTP utilities ───────────────────────────────────────────────────────────
requests>=2.28.0

# ── Flask (only needed in task_hard code snippet, NOT for the env itself) ────
# flask>=3.0.0   ← uncomment only if you import Flask anywhere in server code